numpy更新修正，np.int->int，np.float->float

+++ FEM-python/Advection-Diffusion-python/PrePost.py
@@ -69,8 +69,8 @@ def create_model_json(DataFile):
     model.x = np.array(FEData['x'])
     model.IEN = np.array(FEData['IEN'])
 
-    model.ID  = np.zeros(model.neq,np.int)
-    model.LM  = np.zeros((model.nen,model.nel),np.int)   
+    model.ID  = np.zeros(model.neq,int)
+    model.LM  = np.zeros((model.nen,model.nel),int)   
         
     if 'Exact' in FEData:
         model.Exact = FEData['Exact']
